- fixed bug when input images list is too short for max tracker - fixed bug related to use of maxout
